+++ deploy/scripts/src/mobile-id-check-async/utils/config.ts
@@ -14,6 +14,5 @@ export const config = {
   readidMockApiUrl: getEnv(`MOBILE_ID_CHECK_ASYNC_${environment}_READ_ID_MOCK_API_URL`),
   sessionsApiUrl: getEnv(`MOBILE_ID_CHECK_ASYNC_${environment}_SESSIONS_API_URL`),
   stsMockApiUrl: getEnv(`MOBILE_ID_CHECK_ASYNC_${environment}_STS_MOCK_API_URL`),
-  useProxyApi: getEnv(`MOBILE_ID_CHECK_ASYNC_${environment}_USE_PROXY_API`),
-  biometricSessionTestData: getEnv(`MOBILE_ID_CHECK_ASYNC_${environment}_BIOMETRIC_SESSION_TEST_DATA`)
+  useProxyApi: getEnv(`MOBILE_ID_CHECK_ASYNC_${environment}_USE_PROXY_API`)
 }
